+++ src/main/scala/io/viash/helpers/circe/RichJson.scala
@@ -132,7 +132,7 @@ class RichJson(json: Json) {
    * If an object has a field named "__merge__", that file will be read
    * and be deep-merged with the object itself.
    */
-  def inherit(uri: URI, projectDir: Option[URI], stripInherits: Boolean = true): Json = {
+  def inherit(uri: URI, packageDir: Option[URI], stripInherits: Boolean = true): Json = {
     json match {
       case x if x.isObject =>
         val obj1 = x.asObject.get
@@ -170,7 +170,7 @@ class RichJson(json: Json) {
               if (str == ".") {
                 None
               } else if (str.startsWith("/")) {
-                Some(IO.resolveProjectPath(str, projectDir))
+                Some(IO.resolvePackagePath(str, packageDir))
               } else {
                 Some(uri.resolve(str))
               }
@@ -204,7 +204,7 @@ class RichJson(json: Json) {
                 val newJson1 = Convert.textToJson(str, newURI.toString)
 
                 // recurse through new json as well
-                val newJson2 = newJson1.inherit(newURI, projectDir = projectDir, stripInherits = stripInherits)
+                val newJson2 = newJson1.inherit(newURI, packageDir = packageDir, stripInherits = stripInherits)
 
                 newJson2
             }
@@ -219,11 +219,11 @@ class RichJson(json: Json) {
             throw new ConfigParserMergeException(uri.toString, "invalid merge tag type. Must be a String or Array of Strings", j.toString())
           case None => obj1
         }
-        val obj3 = obj2.mapValues(x => x.inherit(uri, projectDir = projectDir, stripInherits = stripInherits))
+        val obj3 = obj2.mapValues(x => x.inherit(uri, packageDir = packageDir, stripInherits = stripInherits))
         Json.fromJsonObject(obj3)
       case x if x.isArray => 
         val arr1 = x.asArray.get
-        val arr2 = arr1.map(y => y.inherit(uri, projectDir = projectDir, stripInherits = stripInherits))
+        val arr2 = arr1.map(y => y.inherit(uri, packageDir = packageDir, stripInherits = stripInherits))
         Json.fromValues(arr2)
       case _ => json
     }