+++ src/main/scala/io/viash/schemas/JsonSchema.scala
@@ -221,6 +221,7 @@ object JsonSchema {
       !(
         p.`type`.startsWith("Option[") ||
         p.default.isDefined ||
+        p.hasUndocumented ||
         (p.name == "type" && thisParameter.`type` == "PlainFile") // Custom exception, file resources are "kind of" default
       ))
     val requiredJson = required.map(p => Json.fromString(p.name))